+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 + [x] Select: add prop `data-clear-button` (default: false)
 + [x] Grid: optimise styles
 + [x] Hint: fix remove on leave element
++ [x] TreeView: add attribute `data-show-child-count`
 
 ### 4.2.49
 + [x] Select: extended to interop with role=dropdown